requirements.txt
pycryptodome
mysql-connector-python
requests
xmlschema
zeep